+++ NEWS.md
@@ -1,4 +1,4 @@
-# evaluate (development version)
+# evaluate 0.24.0
 
 * The `source` output handler can now take two arguments (the unparsed `src` 
   and the parsed `call`) and choose to affect the displayed source.